Simply arch
Refactor code to simply the arch
